Add files via upload
This was a very interesting project. Simple in concept, a little more complex in execution. Learned quite a bit in the process!

compsci_club_exercise.py: An alternative implementation using the build in chr() and ord() functions of python; a naive implementation is really simple, but doesn't account for control characters and white space winding up being used with the entered offset value, so I had to integrate some error checking and that was tricky.

TODO: error check decryption process for values that yield unprintable results... which we shouldn't have, because the encryption process prevents them from being created in the first place, but we should check for anyway.

check_value.py: just runs through the entire Unicode character space and flags everything you wouldn't want to use in output... the result set is actually pretty small.